requirements.txt
# FastAPI y servidor ASGI
fastapi==0.111.1
uvicorn[standard]==0.30.3
gunicorn==22.0.0

# Pydantic y configuración
pydantic==2.11.4
pydantic-settings==2.9.1
python-dotenv==1.1.0

# Azure SDK (para Microsoft Graph, Azure Management, etc.)
azure-identity==1.17.0
azure-core>=1.34.0 # Versión mínima, permite actualizaciones menores

# Google Ads SDK
google-ads>=25.0.0 # Permite actualizaciones a versiones 25.x.x o superiores
# Las siguientes son dependencias transitivas de google-ads, pero es bueno tenerlas si se especificaron antes
google-auth-oauthlib>=0.5.1
google-api-python-client>=2.0.0 
protobuf>=3.19.0 # Google Ads usa protobuf extensivamente

# Facebook Business SDK (para Meta Ads)
facebook-business>=19.0.0 # Permite actualizaciones

# HTTP y utilidades generales
requests==2.32.3
PyYAML>=5.4 # A menudo útil para configuraciones o requerida por otros SDKs

# Dependencias de criptografía y MSAL (a menudo requeridas por azure-identity u otros flujos de auth)
# Estas versiones son las que tenías, si causan problemas, se pueden flexibilizar o actualizar.
cryptography==43.0.3
cffi>=1.14
pycparser>=2.22
msal>=1.24.0
msal-extensions>=0.3.0
PyJWT[crypto]<3,>=1.0.0 # La restricción <3 es importante para PyJWT

# Typing helpers (a menudo dependencias de Pydantic o FastAPI)
typing-extensions>=4.12.2
typing-inspection>=0.4.0
annotated-types==0.7.0

# Otras dependencias comunes de 'requests' y red
certifi>=2017.4.17
charset-normalizer<4,>=2
idna<4,>=2.5
six>=1.11.0
urllib3<3,>=1.21.1
httpx
# Si usas TikTok Ads SDK explícitamente (aunque no lo modificamos, estaba en tu original)
# tiktok-business-api-sdk>=0.2.6